Fixed damage mitigation algorithm. Adjusted Structure Death test to account for damage mitigation

+++ src/main/game/entities/units/Unit.java
@@ -100,12 +100,13 @@ public abstract class Unit extends Entity implements iAttacker, iDefender, iMove
     /* iMoveable */
     public int getMoveDistance(){ return stats.getSpeed(); }
 
-    /* Stat-adjusted damage taking*/
+    /* Stat-adjusted damage taking */
     @Override
     public void takeDamage(double damage){
-        int armor = stats.getArmor();
+        double armor = stats.getArmor();
         double damageX = 10/(10+armor);
         double adjDamage = damage * damageX;
+
         this.health -= adjDamage;
         this.healthPercent.updateHealthPercentage((double)this.health);
